Create FileProcessing.java
Project that I picked is  - My Search Engine. I have just slowly started implementing some code. I have opened a random file and used the readFile method to read all the contents in the file. This is still very basic and is just for me to get used to working on the basic elements in this project.