Added the TopicModellerGUI object in this class
I added the TopicModellerGUI object in this class. This means that I can access the functions from the TopicModellerGUI class using the object from the WorkWithFiles class. This was primarily done to pass the two maps that were filled with words from the files to the TopicModellerGUI class so that the maps can be be shown in the GUI window.