Fully implemented the FileChooser
I implemented the FileChooser so that the user can add the files that they want to use. A function called receivingMaps() was deleted because I decided to use the actionPerfomed() function to add some of the code from that function and I also added more code. This is the button that asks the user to choose the file. I also have my file object here so I can call the WorkWithFiles class and receive information from that class into my GUI class. This is useful for displaying results in the GUI, which I needed.