Update WorkWithFiles.java
Make quite a lot of changes in this class. I have two functions ( returnFirstMap() and returnSecondMap() ) which are used to sort the map by value in descending order. After this, the list is returned to the GUI class. In the function checkWhetherDocumentsAreCommon(), I used a nested for loop so I can compare each value from the first list with each value in the second list. If there was an occurrence, the occurrences variable was incremented and there was a break command that stopped the inner loop and jumped up to the outer for loop. Also changed the way you calculate the result. I took the top 10 words from both lists and divided the occurrences by 10 and multiplied by 100 to get a percentage. Note: The variable names in all classes will be changed. This was just used to get the program working the way it should.